Refactor | CAKK-81 | Vo 관련 패키지 수정
