+++ cakk-api/src/main/java/com/cakk/api/provider/oauth/PublicKeyProvider.java
@@ -13,9 +13,9 @@ import java.util.Map;
 
 import org.springframework.stereotype.Component;
 
-import com.cakk.client.vo.OidcPublicKey;
-import com.cakk.client.vo.OidcPublicKeyList;
 import com.cakk.common.exception.CakkException;
+import com.cakk.external.vo.key.OidcPublicKey;
+import com.cakk.external.vo.key.OidcPublicKeyList;
 
 @Component
 public class PublicKeyProvider {
@@ -27,13 +27,13 @@ public class PublicKeyProvider {
 	}
 
 	private PublicKey getPublicKey(final OidcPublicKey publicKey) {
-		final byte[] nBytes = decodeBase64(publicKey.n());
-		final byte[] eBytes = decodeBase64(publicKey.e());
+		final byte[] nBytes = decodeBase64(publicKey.getN());
+		final byte[] eBytes = decodeBase64(publicKey.getE());
 
 		final RSAPublicKeySpec publicKeySpec = new RSAPublicKeySpec(new BigInteger(1, nBytes), new BigInteger(1, eBytes));
 
 		try {
-			return KeyFactory.getInstance(publicKey.kty()).generatePublic(publicKeySpec);
+			return KeyFactory.getInstance(publicKey.getKty()).generatePublic(publicKeySpec);
 		} catch (NoSuchAlgorithmException | InvalidKeySpecException e) {
 			throw new CakkException(EXTERNAL_SERVER_ERROR);
 		}
