+++ cakk-api/src/test/java/com/cakk/api/provider/AppleAuthProviderTest.java
@@ -18,10 +18,10 @@ import com.cakk.api.common.base.MockitoTest;
 import com.cakk.api.provider.jwt.JwtProvider;
 import com.cakk.api.provider.oauth.PublicKeyProvider;
 import com.cakk.api.provider.oauth.impl.AppleAuthProvider;
-import com.cakk.client.vo.OidcPublicKeyList;
-import com.cakk.client.web.AppleAuthClient;
 import com.cakk.common.enums.ReturnCode;
 import com.cakk.common.exception.CakkException;
+import com.cakk.external.client.AppleAuthClient;
+import com.cakk.external.vo.key.OidcPublicKeyList;
 
 public class AppleAuthProviderTest extends MockitoTest {
