+++ cakk-external/src/main/kotlin/com/cakk/external/config/S3Config.kt
