+++ cakk-external/src/main/kotlin/com/cakk/external/config/SlackWebhookConfig.kt
