+++ cakk-external/src/main/kotlin/com/cakk/external/extractor/SlackMessageExtractor.kt
