+++ cakk-external/src/main/kotlin/com/cakk/external/template/MessageTemplate.kt
