+++ cakk-external/src/main/kotlin/com/cakk/external/vo/message/ErrorAlertMessage.kt
@@ -1,4 +1,4 @@
-package com.cakk.external.vo
+package com.cakk.external.vo.message
 
 data class ErrorAlertMessage(
 	val serverProfile: String,
