+++ cakk-external/src/main/kotlin/com/cakk/external/vo/message/VerificationMessage.kt
@@ -1,4 +1,4 @@
-package com.cakk.external.vo
+package com.cakk.external.vo.message
 
 data class VerificationMessage(
 	val receiver: String,
