+++ cakk-external/src/main/kotlin/com/cakk/external/vo/s3/PresignedUrl.kt
@@ -1,4 +1,4 @@
-package com.cakk.external.vo
+package com.cakk.external.vo.s3
 
 data class PresignedUrl(
 	val imagePath: String,
